- adds support for payload properties generation openapi C#
